refactor(scheduler): removed task generic
